Move ecosystem screen into ecosystem feature
