+++ apps/mobile-wallet/src/features/ecosystem/EcosystemScreen.tsx
